0.3-dev  (#99)
Core dev for 0.3 ... includes many breaking changes:

* moves csnd plugin / integration tests downstream (https://github.com/bsnacks000/csound-plugins)
* adds license + better docs + doxygen 
* removes `ftable` API and flattens buf gens into their own headers 
* changes APIs for xfade oscils / tabreads to use matrix as its deck
* adds tooling for wavetable interpolation (via matrix)   
* fixes/updates in conv module API 
* better test coverage and various bug fixes

+++ include/dsp/lag.h
@@ -1,7 +1,10 @@
 /**
+ * @file lag.h
  * @brief lag - smoothing filter
- *  - based sc3 Lag class
+ *  - implementation based sc3 Lag class
  */
+
+// SPDX-License-Identifier: MIT
 
 #ifndef DSP_LAG_H
 #define DSP_LAG_H
@@ -18,7 +21,7 @@ extern "C" {
 typedef struct {
     float t_sec, sr;
     // private
-    float alpha_, last_;
+    double alpha_, last_;
 } lag;
 
 /**
